+++ src/test/java/subway/application/line/FindShortestPathServiceTest.java
@@ -11,8 +11,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import subway.application.path.service.FindShortestPathService;
-import subway.domain.fare.DistanceProportionFarePolicy;
-import subway.domain.fare.TotalDistanceFareCalculator;
 import subway.domain.line.Line;
 import subway.domain.line.LineColor;
 import subway.domain.line.LineName;
@@ -38,12 +36,7 @@ class FindShortestPathServiceTest {
 
     @BeforeEach
     void setUp() {
-        final TotalDistanceFareCalculator totalDistanceFareCalculator = new TotalDistanceFareCalculator(List.of(
-                new DistanceProportionFarePolicy(10, 50, 5),
-                new DistanceProportionFarePolicy(50, 1_000_000, 8)
-        ));
         findShortestPathService = new FindShortestPathService(
-                totalDistanceFareCalculator,
                 lineRepository,
                 stationRepository
         );
